SimpleFunctions: Split follow function
